delete folders and files unnecessary, fixed whole pages
