+++ admin/src/app/campaigns/new/page.tsx
@@ -1,13 +1,13 @@
-import CampaignNewFetcher from "@/components/Campaign/CampaignNewFetcher/CampaignNewFetcher";
+import CampaignNew from "@/components/Campaign/CampaignNew";
 
-export default function Page() {
+export default async function Page() {
   return (
     <main className="main">
       <div className="topbar">
         <h1>Create Company</h1>
         {/* <button>Exit</button> */}
       </div>
-      <CampaignNewFetcher />
+      <CampaignNew />
     </main>
   );
 }
