+++ admin/src/components/Campaign/CampaignCreateFetcher.tsx
@@ -1,6 +1,7 @@
 "use server";
 import { postCampaign } from "push-api-client";
 import type { Campaign } from "push-api-client";
+import { getCampaign } from "@/utils/getCampaign";
 
 export async function createCampaignOnServer(data: {
   name: string;
@@ -9,11 +10,7 @@ export async function createCampaignOnServer(data: {
   body: string;
   click_url: string;
 }): Promise<Campaign> {
-  const url = process.env.NEXT_PUBLIC_API_CAMPAIGNS_URL;
-  const token = process.env.PUSH_API_TOKEN;
-
-  if (!url) throw new Error("url is not defined!");
-  if (!token) throw new Error("Token is not defined!");
+  const { token, url } = getCampaign();
 
   return await postCampaign(url, token, data);
 }
