+++ admin/src/components/Campaign/CampaignUpdateFetcher.tsx
@@ -1,6 +1,7 @@
 "use server";
 import { putCompanies } from "push-api-client";
 import type { Campaign } from "push-api-client";
+import { getCampaign } from "@/utils/getCampaign";
 
 console.log(putCompanies);
 
@@ -14,11 +15,7 @@ export async function updateCampaignOnServer(
     click_url: string;
   }
 ): Promise<Campaign> {
-  const url = process.env.NEXT_PUBLIC_API_CAMPAIGNS_URL;
-  const token = process.env.PUSH_API_TOKEN;
-
-  if (!url) throw new Error("url is not defined!");
-  if (!token) throw new Error("Token is not defined!");
+  const { token, url } = getCampaign();
 
   return await putCompanies(url, token, id, data);
 }
