+++ admin/src/components/Campaign/DeleteCampaignOnServer.tsx
@@ -1,8 +1,8 @@
 "use server";
 import { deleteItem } from "push-api-client";
+import { getCampaign } from "@/utils/getCampaign";
 
 export async function DeleteCampaignOnServer(id: string) {
-  const url = process.env.NEXT_PUBLIC_API_CAMPAIGNS_URL!;
-  const token = process.env.PUSH_API_TOKEN!;
+  const { token, url } = getCampaign();
   return await deleteItem(url, token, id);
 }
