+++ admin/src/components/Campaign/RunCampaignOnServer.tsx
@@ -1,8 +1,8 @@
 "use server";
 import { runCampaign } from "push-api-client";
+import { getCampaign } from "@/utils/getCampaign";
 
 export async function RunCampaignOnServer(id: string) {
-  const url = process.env.NEXT_PUBLIC_API_CAMPAIGNS_URL!;
-  const token = process.env.PUSH_API_TOKEN!;
+  const { token, url } = getCampaign();
   return await runCampaign(url, token, id);
 }
