+++ admin/src/components/Devices/DeleteDeviceOnServer.tsx
@@ -1,8 +1,8 @@
 "use server";
 import { deleteItem } from "push-api-client";
+import { getDeviceApiEnv } from "@/utils/getDevice";
 
 export async function DeleteDevicesOnServer(id: string) {
-  const url = process.env.NEXT_PUBLIC_API_DEVICES_URL!;
-  const token = process.env.PUSH_API_TOKEN!;
+  const { token, url } = getDeviceApiEnv();
   return await deleteItem(url, token, id);
 }
