+++ admin/src/components/Devices/DeviceDetailFetcher.tsx
@@ -1,17 +1,10 @@
 import { getDeviceById } from "push-api-client";
 import type { Device } from "push-api-client";
-import DeviceDetail from "@/components/Devices/DeviceDetail/DeviceDetail";
+import DeviceDetail from "@/components/Devices/DeviceDetail";
+import { getDeviceApiEnv } from "@/utils/getDevice";
 
 export default async function DeviceDetailFetcher({ id }: { id: string }) {
-  const url = process.env.NEXT_PUBLIC_API_DEVICES_URL;
-  const token = process.env.PUSH_API_TOKEN;
-
-  if (!url) {
-    return <div>url is not defined!</div>;
-  }
-  if (!token) {
-    return <div>Token is not defined!</div>;
-  }
+  const { token, url } = getDeviceApiEnv();
 
   let device: Device | null = null;
   let error: string | null = null;
